[IMP] mail: search messages with access using SQL
When searching for messages, perform the access checks in SQL.
We can translate rules that have restriction on `model` values directly
into a Domain.

We keep the fallback to search without access rights and filter
afterwards for other cases. We will always limit the number of returned
results from the query to avoid checking the whole table.

The method to `_get_forbidden_access` is flipped to return allowed
records frm a Query instead of self recordset. It allows to reuse it in
the `_search` method so that we don't do an additonal query (one for
search and one for checking access).

Part-of: odoo/odoo#221906
Related: odoo/enterprise#98774
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>